feat!: use Pydantic 2 and move records to use computed_fields
Pydantic 2 has breaking API changes, the most important for us is the
usage of compute_fields.

The accounting system expects datetime fields to be returned as integers
(i.e. epoch). We can use computed_fields for that, using properties.
This way we can internally store datetime objects, and just conver them
to integers when they are accessed. In order to do so, we define _epoch
properties.