Do not render records using .dict() but usin JSOn
The `dict` method of the records are, by design, meant to just conver
the record to a dictionary, without converting data types. If we want to
serialize the records correcty, as we are doing when generating the SSM
records, we need to rely on serialization. See [1] for more context.

[1]: https://github.com/pydantic/pydantic/issues/5155#issuecomment-1460205407

Fixes: #113
SemVer: bugfix